+++ Assets/Photon/Quantum/Runtime/QuantumStats.cs
@@ -1,5 +1,6 @@
 namespace Quantum {
   using Photon.Client;
+  using Quantum.Core;
   using System;
   using System.Diagnostics;
   using UnityEngine;
@@ -20,9 +21,17 @@ namespace Quantum {
     /// </summary>
     public Text FramePredicted;
     /// <summary>
-    /// Number of predicted frames.
+    /// The number of frames currently predicted into the future.
     /// </summary>
     public Text Predicted;
+    /// <summary>
+    /// Verified frames per Unity update.
+    /// </summary>
+    public Text VerifiedFrames;
+    /// <summary>
+    /// Predicted frames per Unity update.
+    /// </summary>
+    public Text PredictedFrames;
     /// <summary>
     /// The last simulation time.
     /// </summary>
@@ -76,8 +85,14 @@ namespace Quantum {
     /// </summary>
     public Text CompactStatsText;
 
+    const float SmoothingTimeSec = 1f;
+
     double _lastTime = 0;
-    int _lastFrame = 0;
+    int _lastVerifiedFrame = 0;
+    SmoothedValue _prediction;
+    SmoothedValue _verifiedFrames;
+    SmoothedValue _predictedFrames;
+    SmoothedValue _simulateTime;
     Stopwatch _networkTimer;
     TrafficStatsSnapshot _snapshotDelta;
 
@@ -106,8 +121,8 @@ namespace Quantum {
         if (QuantumRunner.Default.IsRunning) {
           if (ShowCompactStats) {
             var currentFrame = QuantumRunner.Default.Game.Session.FrameVerified.Number;
-            if (_lastFrame != currentFrame) {
-              _lastFrame = currentFrame;
+            if (_lastVerifiedFrame != currentFrame) {
+              _lastVerifiedFrame = currentFrame;
               var ping = QuantumRunner.Default.Game.Session.Stats.Ping;
               CompactStatsText.text =
                 QuantumRunner.Default.Game.Session.IsOnline
@@ -125,11 +140,27 @@ namespace Quantum {
           if (gameInstance.Session.FramePredicted != null) {
             FrameVerified.text = gameInstance.Session.FrameVerified.Number.ToString();
             FramePredicted.text = gameInstance.Session.FramePredicted.Number.ToString();
+
+            _prediction ??= new SmoothedValue(Mathf.CeilToInt(gameInstance.Session.SessionConfig.UpdateFPS * SmoothingTimeSec));
+            _prediction.Push(gameInstance.Session.FramePredicted.Number - gameInstance.Session.FrameVerified.Number);
+            Predicted.text = ($"{ Math.Round(_prediction.Average, 2):0.00}").ToString();
+
+            _verifiedFrames ??= new SmoothedValue(Mathf.CeilToInt(gameInstance.Session.SessionConfig.UpdateFPS * SmoothingTimeSec));
+            _lastVerifiedFrame = _lastVerifiedFrame < gameInstance.Session.SessionConfig.UpdateFPS ? gameInstance.Session.SessionConfig.UpdateFPS : _lastVerifiedFrame;
+            _verifiedFrames.Push(gameInstance.Session.FrameVerified.Number - _lastVerifiedFrame);
+            _lastVerifiedFrame = gameInstance.Session.FrameVerified.Number;
+            VerifiedFrames.text = $"{Math.Round(_verifiedFrames.Average, 2):0.00} ({_verifiedFrames.Min}-{_verifiedFrames.Max})";
           }
 
-          Predicted.text = gameInstance.Session.PredictedFrames.ToString();
+          _predictedFrames ??= new SmoothedValue(Mathf.CeilToInt(gameInstance.Session.SessionConfig.UpdateFPS * SmoothingTimeSec));
+          _predictedFrames.Push(gameInstance.Session.PredictedFrames);
+          PredictedFrames.text = $"{Math.Round(_predictedFrames.Average, 2):0.00} ({_predictedFrames.Min}-{_predictedFrames.Max})";
+
+          _simulateTime ??= new SmoothedValue(Mathf.CeilToInt(gameInstance.Session.SessionConfig.UpdateFPS * SmoothingTimeSec));
+          _simulateTime.Push((float)gameInstance.Session.Stats.UpdateTime);
+          SimulateTime.text = Math.Round(_simulateTime.Average * 1000, 2) + " ms";
+
           NetworkPing.text = gameInstance.Session.Stats.Ping.ToString();
-          SimulateTime.text = Math.Round(gameInstance.Session.Stats.UpdateTime * 1000, 2) + " ms";
           InputOffset.text = gameInstance.Session.Stats.Offset.ToString();
 
           if (gameInstance.Session.IsStalling) {
@@ -244,6 +275,39 @@ namespace Quantum {
       var num = Math.Round(bytes / Math.Pow(1024, place), 1);
 
       return $"{(Math.Sign(byteCount) * num):0.0} {BytesPerSecondUnits[Math.Min(place, BytesPerSecondUnits.Length - 1)]}";
+    }
+
+    private class SmoothedValue {
+      RingBuffer<float> _ringBuffer;
+      float _avg;
+      float _min;
+      float _max;
+
+      public SmoothedValue(int bufferSize) {
+        _ringBuffer = new RingBuffer<float>(bufferSize);
+      }
+
+      public float Min => _min;
+      public float Max => _max;
+      public float Average => _avg;
+
+      public void Push(float value) {
+        _ringBuffer.PushFront(value);
+
+        var min = float.MaxValue;
+        var max = float.MinValue;
+        var iterator = _ringBuffer.GetIterator();
+        var accumulated = 0.0f;
+        while (iterator.MoveNext()) {
+          accumulated += iterator.Current;
+          min = Mathf.Min(min, iterator.Current);
+          max = Mathf.Max(max, iterator.Current);
+        }
+
+        _avg = accumulated / _ringBuffer.Size;
+        _min = min < float.MaxValue ? min : 0.0f;
+        _max = max > float.MinValue ? max : 0.0f;
+      }
     }
   }
 }